requirements.txt
fastapi
uvicorn
openai-whisper
torch
googletrans==4.0.0-rc1
pydub
edge-tts
ffmpeg-python
python-multipart